Allow signalling server to restart without interrupting audio/video.
A grpc stream can now identify itself with a string id. Identified
streams do not close their peer connections when the grpc connection is
interrupted. A grpc stream that identifies as an existing peer will
reconnect to that peer. Peers are deleted when they go to ICE state
closed.

This allows the signalling server to restart and reconnect to ion-sfu
without interrupting user audio/video.

A grpc stream that does not identify behaves exactly as before.

Identify is actually two things: Reconnect an existing peer, and
Identify a new peer so we can reconnect another time.

If the signalling server restarts we need to reattach to the existing
identified peers. But, if the browser reloads we need to throw away the
existing peers because the PeerConnections are gone (otherwise you get
ErrTransportExists). The signalling server will call Identify before
offer/answer if it thinks it is handling a reconnection. Otherwise it
waits until the connection is established and then identifies the peer.

Possibly this should be two calls, Restore (or Reconnect), and Identify.